+++ testApp/copy/testCreateRequest.cpp
@@ -30,16 +30,27 @@ void testCreateRequest() {
     string request = "";
     if(debug) { cout  << "request " << request <<endl;}
     PVStructurePtr pvRequest = createRequest->createRequest(request);
-    if(pvRequest.get()==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
     if(debug) { cout << *pvRequest << endl;}
     testOk1(pvRequest.get()!=NULL);
     testOk1(pvRequest->getStructure()->getNumberFields()==0);
     testPass("request %s",request.c_str());
 
+    request = "record[]field()getField()putField()";
+    if(debug) { cout  << "request " << request <<endl;}
+    pvRequest = createRequest->createRequest(request);
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(debug) { cout << *pvRequest << endl;}
+    testOk1(pvRequest.get()!=NULL);
+    testOk1(pvRequest->getSubField("field").get()!=NULL);
+    testOk1(pvRequest->getSubField("putField").get()!=NULL);
+    testOk1(pvRequest->getSubField("getField").get()!=NULL);
+    testPass("request %s",request.c_str());
+
     request = "record[a=b,x=y]field(a) putField(a),getField(a)";
     if(debug) { cout  << "request " << request <<endl;}
     pvRequest = createRequest->createRequest(request);
-    if(pvRequest.get()==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
     if(debug) { cout << *pvRequest << endl;}
     testOk1(pvRequest.get()!=NULL);
     pvString = pvRequest->getSubField<PVString>("record._options.a");
@@ -56,7 +67,7 @@ void testCreateRequest() {
     request = "field(a.b[x=y])";
     if(debug) { cout  << "request " << request <<endl;}
     pvRequest = createRequest->createRequest(request);
-    if(pvRequest.get()==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
     if(debug) { cout << *pvRequest << endl;}
     testOk1(pvRequest.get()!=NULL);
     pvString = pvRequest->getSubField<PVString>("field.a.b._options.x");
@@ -67,7 +78,7 @@ void testCreateRequest() {
     request = "field(a.b{c.d})";
     if(debug) { cout  << "request " << request <<endl;}
     pvRequest = createRequest->createRequest(request);
-    if(pvRequest.get()==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
     if(debug) { cout << *pvRequest << endl;}
     testOk1(pvRequest.get()!=NULL);
     testOk1(pvRequest->getSubField("field.a.b.c.d").get()!=NULL);
@@ -76,7 +87,7 @@ void testCreateRequest() {
     request = "field(a.b[x=y]{c.d})";
     if(debug) { cout  << "request " << request <<endl;}
     pvRequest = createRequest->createRequest(request);
-    if(pvRequest.get()==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
     if(debug) { cout << *pvRequest << endl;}
     testOk1(pvRequest.get()!=NULL);
     pvString = pvRequest->getSubField<PVString>("field.a.b._options.x");
@@ -88,7 +99,7 @@ void testCreateRequest() {
     request = "field(a.b[x=y]{c.d[x=y]})";
     if(debug) { cout  << "request " << request <<endl;}
     pvRequest = createRequest->createRequest(request);
-    if(pvRequest.get()==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
     if(debug) { cout << *pvRequest << endl;}
     testOk1(pvRequest.get()!=NULL);
     pvString = pvRequest->getSubField<PVString>("field.a.b._options.x");
@@ -102,7 +113,7 @@ void testCreateRequest() {
     request = "record[a=b,c=d] field(a.a[a=b]{a.a[a=b]},b.a[a=b]{a,b})";
     if(debug) { cout  << "request " << request <<endl;}
     pvRequest = createRequest->createRequest(request);
-    if(pvRequest.get()==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
     if(debug) { cout << *pvRequest << endl;}
     testOk1(pvRequest.get()!=NULL);
     pvString = pvRequest->getSubField<PVString>("field.a.a._options.a");
@@ -122,7 +133,7 @@ void testCreateRequest() {
     request = "alarm,timeStamp,power.value";
     if(debug) { cout  << "request " << request <<endl;}
     pvRequest = createRequest->createRequest(request);
-    if(pvRequest.get()==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
     if(debug) { cout << *pvRequest << endl;}
     testOk1(pvRequest.get()!=NULL);
     testOk1(pvRequest->getSubField("field.alarm").get()!=NULL);
@@ -133,7 +144,7 @@ void testCreateRequest() {
     request = "record[process=true]field(alarm,timeStamp,power.value)";
     if(debug) { cout  << "request " << request <<endl;}
     pvRequest = createRequest->createRequest(request);
-    if(pvRequest.get()==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
     if(debug) { cout << *pvRequest << endl;}
     testOk1(pvRequest.get()!=NULL);
     pvString = pvRequest->getSubField<PVString>("record._options.process");
@@ -147,7 +158,7 @@ void testCreateRequest() {
     request = "record[process=true]field(alarm,timeStamp[algorithm=onChange,causeMonitor=false],power{value,alarm})";
     if(debug) { cout  << "request " << request <<endl;}
     pvRequest = createRequest->createRequest(request);
-    if(pvRequest.get()==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
     if(debug) { cout << *pvRequest << endl;}
     testOk1(pvRequest.get()!=NULL);
     pvString = pvRequest->getSubField<PVString>("record._options.process");
@@ -168,7 +179,7 @@ void testCreateRequest() {
     request = "record[int=2,float=3.14159]field(alarm,timeStamp[shareData=true],power.value)";
     if(debug) { cout  << "request " << request <<endl;}
     pvRequest = createRequest->createRequest(request);
-    if(pvRequest.get()==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
     if(debug) { cout << *pvRequest << endl;}
     testOk1(pvRequest.get()!=NULL);
     pvString = pvRequest->getSubField<PVString>("record._options.int");
@@ -190,7 +201,7 @@ void testCreateRequest() {
     	+ "current{value,alarm},voltage{value,alarm})";
     if(debug) { cout  << "request " << request <<endl;}
     pvRequest = createRequest->createRequest(request);
-    if(pvRequest.get()==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
     if(debug) { cout << *pvRequest << endl;}
     testOk1(pvRequest.get()!=NULL);
     testOk1(pvRequest->getSubField("putField.power.value").get()!=NULL);
@@ -210,7 +221,7 @@ void testCreateRequest() {
         + "})";
     if(debug) { cout  << "request " << request <<endl;}
     pvRequest = createRequest->createRequest(request);
-    if(pvRequest.get()==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
     if(debug) { cout << *pvRequest << endl;}
     testOk1(pvRequest.get()!=NULL);
     testOk1(pvRequest->getSubField("field.alarm").get()!=NULL);
@@ -232,7 +243,7 @@ void testCreateRequest() {
         + ")";
     if(debug) { cout  << "request " << request <<endl;}
     pvRequest = createRequest->createRequest(request);
-    if(pvRequest.get()==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
     if(debug) { cout << *pvRequest << endl;}
     testOk1(pvRequest.get()!=NULL);
     testOk1(pvRequest->getSubField("putField.power.value").get()!=NULL);
@@ -265,7 +276,7 @@ void testCreateRequest() {
     request = "a{b{c{d}}}";
     if(debug) { cout  << "request " << request <<endl;}
     pvRequest = createRequest->createRequest(request);
-    if(pvRequest.get()==NULL) { cout<< createRequest->getMessage() << endl;}
+    if(pvRequest==NULL) { cout<< createRequest->getMessage() << endl;}
     if(debug) { cout << *pvRequest << endl;}
     testOk1(pvRequest.get()!=NULL);
     testOk1(pvRequest->getSubField("field.a.b.c.d").get()!=NULL);
@@ -279,23 +290,23 @@ void testCreateRequest() {
         + "ps1{alarm,timeStamp,power{value,alarm},current{value,alarm},voltage{value,alarm}"
         + ")";
     if(debug) { cout  << "request " << request <<endl;}
-    if(debug) { cout << endl << "Error Expected for next call!!" << endl;}
+    cout << endl << "Error Expected for next call!!" << endl;
     pvRequest = createRequest->createRequest(request);
-    if(debug) {cout << "reason " << createRequest->getMessage() << endl;}
+    cout << "reason " << createRequest->getMessage() << endl;
     testPass("request %s",request.c_str());
 
     request = "record[process=true,power.value";
     if(debug) { cout  << "request " << request <<endl;}
-    if(debug) { cout << endl << "Error Expected for next call!!" << endl;}
+    cout << endl << "Error Expected for next call!!" << endl;
     pvRequest = createRequest->createRequest(request);
-    if(debug) {cout << "reason " << createRequest->getMessage() << endl;}
+    cout << "reason " << createRequest->getMessage() << endl;
     testOk1(pvRequest.get()==NULL);
     testPass("request %s",request.c_str());
 }
 
 MAIN(testCreateRequest)
 {
-    testPlan(112);
+    testPlan(117);
     testCreateRequest();
     return testDone();
 }